Trim silence and cut audio to fixed size
